requirements.txt
anyio==3.6.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
h11==0.14.0
httpcore==0.16.2
httpx==0.23.1
idna==3.4
outcome==1.2.0
psycopg2==2.9.5
pycparser==2.21
PySocks==1.7.1
python-telegram-bot==20.0a6
rfc3986==1.5.0
selenium==4.7.2
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.13
wsproto==1.2.0
